refactor(views): extract shared owned-public-message scaffolding into utils
